feat(fetch): 🎉 integrate with cross-fetch to global for fetch support
