+++ examples/express-ts/paths/_index.ts
@@ -1,0 +1,17 @@
+import { customResolver } from './custom-resolver'
+import { fetchApi } from './fetch-api'
+import { guard } from './guard'
+import { hello } from './hello'
+import { helloName } from './hello/name'
+import { requestValidation } from './request-validation'
+import { validation } from './validation'
+
+export const paths = [
+	customResolver,
+	fetchApi,
+	guard,
+	hello,
+	helloName,
+	requestValidation,
+	validation,
+]
