+++ examples/express-ts/paths/fetch-api.ts
@@ -1,0 +1,17 @@
+import { z } from 'zod'
+import { router } from '../router'
+
+export const fetchApi = router.path('/fetch-api')
+
+fetchApi.handler({
+	method: 'get',
+	response: {
+		200: z.any(),
+	},
+	async resolver() {
+		const result = await fetch('https://pokeapi.co/api/v2/pokemon/ditto')
+		return {
+			200: await result.json(),
+		}
+	},
+})
