+++ examples/express-ts/router.ts
@@ -1,0 +1,4 @@
+import { Router } from 'express'
+import { createRouter } from 'bulbasaur-express'
+
+export const router = createRouter(Router())
