+++ package.json
@@ -1,72 +1,73 @@
 {
-  "name": "bulbasaur-express",
-  "description": "",
-  "version": "0.0.0",
-  "license": "MIT",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "lib"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/wai-lin/bulbasaur-express.git"
-  },
-  "bugs": {
-    "url": "https://github.com/wai-lin/bulbasaur-express/issues"
-  },
-  "homepage": "https://github.com/wai-lin/bulbasaur-express#readme",
-  "keywords": [
-    "express",
-    "express-framework"
-  ],
-  "scripts": {
-    "pre:build": "rm -rf dist",
-    "build": "npm run pre:build && tsc && tsc-alias ",
-    "build:watch": "npm-watch build",
-    "update:pkg": "ncu -i --format group",
-    "commit": "git add -A && cz",
-    "gen:mach-type": "xstate typegen",
-    "gen:mach-type:watch": "xstate typegen --watch"
-  },
-  "watch": {
-    "build": {
-      "patterns": [
-        "lib"
-      ],
-      "extensions": "ts",
-      "quiet": false
-    }
-  },
-  "config": {
-    "commitizen": {
-      "path": "./node_modules/cz-emojii-conventional-changelog"
-    }
-  },
-  "dependencies": {
-    "express": "^4.18.1",
-    "npm-watch": "^0.11.0",
-    "ts-node": "^10.9.1",
-    "tsc-alias": "^1.7.0",
-    "type-fest": "^2.17.0",
-    "typescript": "^4.7.4",
-    "zod": "^3.17.9"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.13",
-    "@types/node": "^18.0.6",
-    "@typescript-eslint/eslint-plugin": "^5.30.7",
-    "@typescript-eslint/parser": "^5.30.7",
-    "commitizen": "^4.2.5",
-    "cz-emojii-conventional-changelog": "^0.0.0-semantically-released.0",
-    "eslint": "^8.20.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-n": "^15.2.4",
-    "eslint-plugin-promise": "^6.0.0",
-    "nodemon": "^2.0.19",
-    "prettier": "^2.7.1"
-  }
+	"name": "bulbasaur-express",
+	"description": "",
+	"version": "0.0.0",
+	"license": "MIT",
+	"main": "dist/index.js",
+	"types": "dist/index.d.ts",
+	"files": [
+		"dist",
+		"lib"
+	],
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/wai-lin/bulbasaur-express.git"
+	},
+	"bugs": {
+		"url": "https://github.com/wai-lin/bulbasaur-express/issues"
+	},
+	"homepage": "https://github.com/wai-lin/bulbasaur-express#readme",
+	"keywords": [
+		"express",
+		"express-framework"
+	],
+	"scripts": {
+		"pre:build": "rm -rf dist",
+		"build": "npm run pre:build && tsc && tsc-alias ",
+		"build:watch": "npm-watch build",
+		"update:pkg": "ncu -i --format group",
+		"commit": "git add -A && cz",
+		"gen:mach-type": "xstate typegen",
+		"gen:mach-type:watch": "xstate typegen --watch"
+	},
+	"watch": {
+		"build": {
+			"patterns": [
+				"lib"
+			],
+			"extensions": "ts",
+			"quiet": false
+		}
+	},
+	"config": {
+		"commitizen": {
+			"path": "./node_modules/cz-emojii-conventional-changelog"
+		}
+	},
+	"dependencies": {
+		"cross-fetch": "^3.1.5",
+		"express": "^4.18.1",
+		"npm-watch": "^0.11.0",
+		"ts-node": "^10.9.1",
+		"tsc-alias": "^1.7.0",
+		"type-fest": "^2.17.0",
+		"typescript": "^4.7.4",
+		"zod": "^3.17.9"
+	},
+	"devDependencies": {
+		"@types/express": "^4.17.13",
+		"@types/node": "^18.0.6",
+		"@typescript-eslint/eslint-plugin": "^5.30.7",
+		"@typescript-eslint/parser": "^5.30.7",
+		"commitizen": "^4.2.5",
+		"cz-emojii-conventional-changelog": "^0.0.0-semantically-released.0",
+		"eslint": "^8.20.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-config-standard": "^17.0.0",
+		"eslint-plugin-import": "^2.26.0",
+		"eslint-plugin-n": "^15.2.4",
+		"eslint-plugin-promise": "^6.0.0",
+		"nodemon": "^2.0.19",
+		"prettier": "^2.7.1"
+	}
 }
